Improve update
The original integration has an ugly management of updates, since no coordinator was prepared. 
Implementing a coordinator now would mean rewrite a big part of the cose, so i found a partial workarount triggering an update every time the climate entities do something, delayed of a couple of seconds (time for the cube to make new values available).